requirements.txt
flask>=2.2.0
flask-cors>=3.0.10
websocket-client>=1.5.2
requests>=2.31.0
python-dotenv>=1.0.1
gunicorn>=21.2.0